宠物服务--宠物管理，三表之间简单的crud

+++ pet-shop-pet/src/main/java/com/litchi/petshop/pet/controller/PetBreedController.java
@@ -1,8 +1,9 @@
 package com.litchi.petshop.pet.controller;
 
-import java.util.Arrays;
-import java.util.Map;
+import java.util.*;
 
+import com.litchi.petshop.pet.entity.PetEntity;
+import com.litchi.petshop.pet.service.PetService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestBody;
@@ -30,6 +31,9 @@ public class PetBreedController {
     @Autowired
     private PetBreedService petBreedService;
 
+    @Autowired
+    PetService petService;
+
     /**
      * 列表
      */
@@ -41,6 +45,12 @@ public class PetBreedController {
         return R.ok().put("page", page);
     }
 
+    @RequestMapping("/listBreed")
+    public R listBreed(@RequestParam Map<String, Object> params){
+        PageUtils page = petBreedService.queryBreedPage(params);
+
+        return R.ok().put("page", page);
+    }
 
     /**
      * 信息
@@ -81,8 +91,35 @@ public class PetBreedController {
     @RequestMapping("/delete")
     //@RequiresPermissions("pet:petbreed:delete")
     public R delete(@RequestBody Integer[] breedIds){
-		petBreedService.removeByIds(Arrays.asList(breedIds));
 
+        // 查看毛色是否被pet关联到，被关联到则无法删除
+        List<PetEntity> petEntityList = petService.list();
+        //所有被关联到的breedId
+        Set<Integer> relatedAllIds = new HashSet<>();
+        for (PetEntity petEntity : petEntityList) {
+            relatedAllIds.add(petEntity.getBreedId());
+        }
+
+        // 没被关联的colorId
+        List<Integer> resultIds = new ArrayList<>();
+        // 被关联的colorId
+        List<Integer> relatedIds = new ArrayList<>();
+
+        for (Integer breedId : breedIds) {
+            if (!relatedAllIds.contains(breedId)) {
+                resultIds.add(breedId);
+            } else {
+                //被关联到的breedId
+                relatedIds.add(breedId);
+            }
+        }
+
+        if (relatedIds.size() != 0) {
+            return R.error().put("msg", "编号为：" + Arrays.toString(relatedIds.toArray()) + "被pet关联，无法删除");
+        }
+        if (resultIds.size() != 0) {
+            petBreedService.removeByIds(Arrays.asList(breedIds));
+        }
         return R.ok();
     }
